feat(tolling): implement batched sum_fg_for_participants and enhance precision in calculations
- Added `sum_fg_for_participants_batch` method to `TollingDistributionRepository` for efficient aggregation of gross, fee, and net amounts across multiple tolling lots.
- Updated `_calculate_lines` in `TollingService` to ensure all monetary values are handled as `Decimal` to prevent floating-point inaccuracies.
- Introduced concurrency control with advisory locks in `lock_balance_key` to prevent race conditions during stock balance updates.
- Implemented period locking in `assert_open_period` to block transactions on closed reporting periods.
- Added tests to verify decimal precision and concurrency behavior, ensuring correct handling of stock adjustments and tolling distributions.
- Created a unique index to enforce a single open tolling lot per company, preventing multiple open lots.
- Enhanced integration tests to cover scenarios for negative stock warnings and double counting of finished goods in daily reports.

+++ backend/app/modules/adjustment/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import (
@@ -74,8 +75,8 @@ class InventoryAdjustmentLine(Base, UUIDPrimaryKeyMixin):
     waste_type: Mapped[WasteType | None] = mapped_column(nullable=True)
     pkg_item_type: Mapped[PackagingItemType | None] = mapped_column(nullable=True)
 
-    quantity_kg_before: Mapped[float] = mapped_column(Numeric(15, 3), nullable=False, default=0)
-    quantity_kg_after: Mapped[float] = mapped_column(Numeric(15, 3), nullable=False, default=0)
+    quantity_kg_before: Mapped[Decimal] = mapped_column(Numeric(15, 3), nullable=False, default=0)
+    quantity_kg_after: Mapped[Decimal] = mapped_column(Numeric(15, 3), nullable=False, default=0)
     bags_before: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     bags_after: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     units_before: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
